[REFACTOR] Plant and search controller
