+++ server/controllers/Plant/Index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  get: require('./Get.js'),
+  post: require('./Post.js'),
+  patch: require('./Patch.js'),
+  delete: require('./Delete.js'),
+};
