+++ server/routes/Plants.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const plantController = require('../controllers/PlantController');
+const plantController = require('../controllers/Plant/Index');
 const interiorController = require('../controllers/Interior/Index');
 const isAuth = require('../middlewares/Authentication');
 const plantImageHandler = require('../middlewares/PlantImageHandler.js');
