--- server/routes/Search.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const searchController = require('../controllers/SearchController');
-
-router.get('/', searchController.get);
-
-module.exports = router;
